[VM] * Removed all references to the various C++ iostream classes; they bloat the code, the templates take forever to compile, and seems (according to private tests I won't publish here) to be slower than C-like functions * As a result of the above, reimplemented ModuleReader to use Windows API calls instead of C++ fstreams. The Windows API offers three advantages:     * (Somewhat) easier to work with     * More detailed error information     * Wide-character path names (fopen has no wide-character equivalent, sadly) (If Ovum is ever ported to other systems, we will introduce #defines and #ifs to switch to a POSIX API outside of Windows, or something else that works. (But probably not C++ fstreams.) * Moved the implementation of ModuleReader to module.cpp; only a forward declaration is in ov_module.internal.h now. No other file needs to use ModuleReader. * Disabled security checks in Release mode, as it marginally slows down sensitive code (such as PushStackFrame, InvokeMethodOverload and Evaluate) * Improved initialization of various members (pointers in the Member classes and StringHash<T>, for example), to counteract destructor bugs (can't delete pointer 0xcccccccc!) * Changed LocalOffset to be a pre-multiplied (by sizeof(Value)) int32_t, as it vastly improves execution speed * Changed argument counts in methods, both in the public C API and internally, to use type uint32_t instead of unsigned int or uint16_t * Added partial implementation of threaded bytecode execution (to be described later; #define THREADED_EXECUTION to enable it, and use gcc (don't enable yet)); not supported in the MSVC compiler * Added several new intermediate branching instructions for certain common situations: breq, brneq, brlt, brgt, brlte, brgte. These are produced when an eq, lt, gt, lte or gte precedes a brtrue or brfalse, and there are no incoming branches to the branch instruction. * Added separate implementations of (almost) all the instructions, removing code like "frame->stackCount += opc & 1", which was found to marginally slow down execution * Removed the StackChanges array from thread.opcodes.cpp; it didn't actually add anything useful
[aves]
* Removed all references to <iostream>, for reasons explained above
* Rewrote aves_readLine in terms of Windows API functions; when porting to other systems, there will be a preprocessor condition that selects the right implementation
* Added (a slightly modified version of) the fantastic dtoa.c library, by Daniel M. Gay, which formats doubles into strings, and parses strings into doubles
* Rewrote aves_Real_toString with dtoa.c functions
* Added toString override to aves.{Stack,Set,Queue}

[Ovum]
* Removed the _tout macro (was only used in one place)
* References to <iostream> will be removed in the next commit, probably

+++ VM/ov_thread.h
@@ -16,11 +16,11 @@ OVUM_API void VM_PushString(ThreadHandle thread, String *str);
 
 // Pops a single value off the top of the evaluation stack.
 OVUM_API Value VM_Pop(ThreadHandle thread);
-OVUM_API void VM_PopN(ThreadHandle thread, const unsigned int n);
+OVUM_API void VM_PopN(ThreadHandle thread, const uint32_t n);
 
 OVUM_API void VM_Dup(ThreadHandle thread);
 
-OVUM_API Value *VM_Local(ThreadHandle thread, const unsigned int n);
+OVUM_API Value *VM_Local(ThreadHandle thread, const uint32_t n);
 
 // Invokes a value on the evaluation stack.
 // If S[0] is the top value on the stack, then S[argCount] is the value that is invoked.
@@ -30,7 +30,7 @@ OVUM_API Value *VM_Local(ThreadHandle thread, const unsigned int n);
 //     The number of arguments passed in the invocation, NOT including the instance.
 //   result:
 //     The return value of the invocation. If null, the return value is pushed onto the stack.
-OVUM_API void VM_Invoke(ThreadHandle thread, const unsigned int argCount, Value *result);
+OVUM_API void VM_Invoke(ThreadHandle thread, const uint32_t argCount, Value *result);
 
 // Invokes a member of a value on the evaluation stack.
 // If S[0] is the top value on the stack, then S[argCount] is the value whose member is invoked.
@@ -43,7 +43,7 @@ OVUM_API void VM_Invoke(ThreadHandle thread, const unsigned int argCount, Value 
 //     The number of arguments to pass to the method, NOT including the instance.
 //   result:
 //     The return value of the invocation. If null, the return value is pushed onto the stack.
-OVUM_API void VM_InvokeMember(ThreadHandle thread, String *name, const unsigned int argCount, Value *result);
+OVUM_API void VM_InvokeMember(ThreadHandle thread, String *name, const uint32_t argCount, Value *result);
 
 // Invokes a specific method with arguments from the evaluation stack.
 //   thread:
@@ -54,7 +54,7 @@ OVUM_API void VM_InvokeMember(ThreadHandle thread, String *name, const unsigned 
 //     The number of arguments to pass to the method, NOT including the instance.
 //   result:
 //     The return value of the invocation. If null, the return value is pushed onto the stack.
-OVUM_API void VM_InvokeMethod(ThreadHandle thread, MethodHandle method, const unsigned int argCount, Value *result);
+OVUM_API void VM_InvokeMethod(ThreadHandle thread, MethodHandle method, const uint32_t argCount, Value *result);
 // Invokes an operator on one or two values on the evaluation stack.
 //   thread:
 //     The thread on which to invoke the operator.
@@ -108,13 +108,13 @@ OVUM_API void VM_StoreMember(ThreadHandle thread, String *member);
 //     The number of arguments to invoke the indexer with, not including the instance.
 //   result:
 //     The result of loading the member. If null, this value is pushed onto the stack.
-OVUM_API void VM_LoadIndexer(ThreadHandle thread, const uint16_t argCount, Value *result);
+OVUM_API void VM_LoadIndexer(ThreadHandle thread, const uint32_t argCount, Value *result);
 // Stores the top value on the stack into the indexer of the second stack value.
 //   thread:
 //     The thread on which to execute the indexer setter.
 //   argCount:
 //     The number of arguments to invoke the indexer with, not including the instance or value.
-OVUM_API void VM_StoreIndexer(ThreadHandle thread, const uint16_t argCount);
+OVUM_API void VM_StoreIndexer(ThreadHandle thread, const uint32_t argCount);
 
 // Loads the value of the specified static field.
 //   thread:
